refactor: removed custom exceptions

+++ src/main/java/org/example/model/dto/CompletedCourseCreateRequest.java
@@ -7,8 +7,8 @@ import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 
 public record CompletedCourseCreateRequest(
-        @Min(1) int studentId,
-        @Min(1) int courseId,
+        @Min(1) Long studentId,
+        @Min(1) Long courseId,
         @NotNull @Min(1) @Max(5) int grade,
         @NotNull LocalDate completionDate
 ) {}
